[#12] refactor: 게임 정보 더 보기 버튼 컴포넌트 분리
